chg: add some LinOTPAuth fixes and documentation
The base functionality to authenticate against LinOTP is already present
for some time. This change introduces some bug fixes and quality of life
changes (documentation).  This also introduces the standard example line
in the `bootstrap.default.php`.

The standard `AppComponent::beforeFilter` execution
sometimes triggers more then one authentication attempt. In the case of
invalid (or missing additional) credentials LinOTP would sometimes
increment the fail counter of a user twice. This has been addressed by
caching the result of the authentication after the first lookup during
the current request.

Also we are now retaining the LinOTPAuth configuration in your settings
after you have modified the settings in the MISP Admin UI.

The `detail` property of LinOTP respones as previously not extracted
properly. This has been fixed.

The content type of LinOTP respones was previously only checked for an
exact match against "application/json". It is common for webservers to
inject an encoding (and other information) to the content type. This has
been fixed and should now work with any standard LinOTP installation.