new: MISPAuthComponent to handle dynamic loginAction
Since we must be able to override the loginAction based on the enabled
plugins we need a custom AuthComponent.

Whenever the LinOTP Auth plugin was selected for authentication
redirects to the login page should go there. This does not yet modify
the behaviour of new logins but ensures that we have a consistent login
logic as soon as the initial logins are being redirected there as well.