add: [LinOTPAuth] add initial support for a multi-factor login
Since plugins that provide views and Controllers are required to
implement a Models and AppController. Those have been added in a very
bare version that serves our purposes.

This also adds all the interaction between the custom view and the auth
component. The assumption is that we are always using both together.
Using the view without the auth part is not supported and will fail all
authentications.

For the continued authentication to work (e.g. a user provides another
credentail after the initial username / password step) we have to start
passing a transaction id between the requests and to LinOTP. Those
changes have been made. The process of passing and storing the
information isn't super elegant but it should be safe as far as I can
tell.